+++ apps/cowswap-frontend/src/modules/swap/state/swapRawStateAtom.ts
@@ -5,12 +5,8 @@ import { atomWithPartialUpdate } from '@cowprotocol/common-utils'
 import { getJotaiIsolatedStorage } from '@cowprotocol/core'
 import { OrderKind, SupportedChainId } from '@cowprotocol/cow-sdk'
 
-import {
-  DEFAULT_TRADE_DERIVED_STATE,
-  ExtendedTradeRawState,
-  getDefaultTradeRawState,
-  TradeDerivedState,
-} from 'modules/trade'
+import { DEFAULT_TRADE_DERIVED_STATE, TradeDerivedState } from 'modules/trade/types/TradeDerivedState'
+import { ExtendedTradeRawState, getDefaultTradeRawState } from 'modules/trade/types/TradeRawState'
 
 export interface SwapDerivedState extends TradeDerivedState {
   isUnlocked: boolean