+++ apps/cowswap-frontend/src/modules/twap/updaters/QuoteObserverUpdater.tsx
@@ -1,5 +1,5 @@
 import { useAtomValue } from 'jotai'
-import { useMemo } from 'react'
+import { useEffect, useMemo } from 'react'
 
 import { usePrevious } from '@cowprotocol/common-hooks'
 
@@ -40,7 +40,7 @@ export function QuoteObserverUpdater(): null {
     return adjustedForParts.quotient.toString()
   }, [isLoading, numberOfPartsValue, buyAmount, prevNumberOfParts])
 
-  useMemo(() => {
+  useEffect(() => {
     if (!outputCurrency || !quote || !quoteBuyAmount) {
       return
     }