+++ libs/common-const/src/getAvailableChainsText.ts
@@ -4,8 +4,8 @@ import { SORTED_CHAIN_IDS } from './chainInfo'
 
 export function getAvailableChainsText(): string {
   return SORTED_CHAIN_IDS.reduce((acc, chainId) => {
-    const { label, isTestnet, isUnderDevelopment } = ALL_SUPPORTED_CHAINS_MAP[chainId]
-    if (!isUnderDevelopment) {
+    const { label, isTestnet, isUnderDevelopment, isDeprecated } = ALL_SUPPORTED_CHAINS_MAP[chainId]
+    if (!isUnderDevelopment && !isDeprecated) {
       acc.push(`${label}${isTestnet ? ' (testnet)' : ''}`)
     }
     return acc