+++ libs/wallet-provider/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [2.1.3](https://github.com/cowprotocol/cowswap/compare/wallet-provider-v2.1.2...wallet-provider-v2.1.3) (2026-02-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 2.2.2
+
+## [2.1.2](https://github.com/cowprotocol/cowswap/compare/wallet-provider-v2.1.1...wallet-provider-v2.1.2) (2026-02-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 2.2.1
+
 ## [2.1.1](https://github.com/cowprotocol/cowswap/compare/wallet-provider-v2.1.0...wallet-provider-v2.1.1) (2026-02-13)